changed Quote Volume data type to decimal (#113)
* Changed Volume long to decimal

Indicators.Volume type changed
ObvResult would have needed to be changed to match, didn't want to take that step since it was oiut of the scope of the issue so I cast Volume back to long

* Updated Documentation to reflect datatype changes

GUIDE.md in root folder now reflects cahnge of long to decimal

README.md in Obv Now relfects change of of obv data type from long to decimal (as well as Obv property in ObvResult)

+++ Indicators/Obv/Obv.Models.cs
@@ -5,6 +5,6 @@ namespace Skender.Stock.Indicators
     [Serializable]
     public class ObvResult : ResultBase
     {
-        public long Obv { get; set; }
+        public decimal Obv { get; set; }
     }
 }
